+++ neuracore/core/const.py
@@ -7,13 +7,17 @@ import os
 REMOTE_RECORDING_TRIGGER_ENABLED = (
     os.getenv("NEURACORE_REMOTE_RECORDING_TRIGGER_ENABLED", "True").lower() == "true"
 )
-LIVE_DATA_ENABLED = os.getenv("NEURACORE_LIVE_DATA_ENABLED", "True").lower() == "true"
+PROVIDE_LIVE_DATA = os.getenv("NEURACORE_PROVIDE_LIVE_DATA", "True").lower() == "true"
+CONSUME_LIVE_DATA = os.getenv("NEURACORE_CONSUME_LIVE_DATA", "True").lower() == "true"
+
 
 API_URL = os.getenv("NEURACORE_API_URL", "https://api.neuracore.app/api")
 MAX_DATA_STREAMS = 50
-
-
 MAX_INPUT_ATTEMPTS = 3
+
+STREAMING_MINIMUM_BACKOFF_TIME_S = 0.05
+STREAMING_MAXIMUM_BACKOFF_TIME_S = 5
+
 CONFIRMATION_INPUT = {
     "yes",
     "y",